🐛 Fix DDSP cascade for batched [B,C,T] input + STFT window dtype
Address Copilot review on #103:
- BiquadFunction.forward/backward flatten any leading dims to [N, T] for
  the native kernel (which TORCH_CHECKs x.dim()==2) and restore the shape,
  matching the convention used by every other torchfx._ops dispatcher. The
  docstring already advertised [B, C, T]; it now actually works.
- multires_stft_loss builds the Hann window with dtype=x.dtype to avoid a
  latent float32->float64 promotion inside torch.stft.
- Add a [B, C, T] regression test (per-channel parity + grad flow).

+++ src/torchfx/ddsp/losses.py
@@ -27,7 +27,7 @@ def multires_stft_loss(x: Tensor, y: Tensor, sizes: tuple[int, ...] = (512, 1024
     """
     loss = x.new_zeros(())
     for n_fft in sizes:
-        win = torch.hann_window(n_fft, device=x.device)
+        win = torch.hann_window(n_fft, device=x.device, dtype=x.dtype)
         x_mag = torch.stft(x, n_fft, n_fft // 4, window=win, return_complex=True).abs()
         y_mag = torch.stft(y, n_fft, n_fft // 4, window=win, return_complex=True).abs()
         loss = loss + torch.norm(x_mag - y_mag) / (torch.norm(y_mag) + 1e-8)
